+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - `ISOCurrencyProvider::getCurrency()`
 - **`RationalMoney` is now always simplified to lowest terms:** `USD 25/100` is automatically simplified to `USD 1/4`
 - `Currency::is()` has been removed, use `Currency::isEqualTo()` instead
+- `MoneyBag::getAmount()` has been removed, use `MoneyBag::getMoney()->getAmount()` instead
 
 ⚠️ **Deprecations**